requirements.txt
aiofiles==23.2.1
amqp==5.1.1
anyio==3.7.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
ecdsa==0.18.0
fastapi==0.97.0
h11==0.14.0
idna==3.4
kombu==5.3.1
MarkupSafe==2.1.3
prompt-toolkit==3.0.39
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.12
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
redis==5.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.6
